Added a button to Regenerate the Data Model in the Data Model Browser
The Data Model is currently stored inside of Blender's properties as a
single string containing a Python "pickle" of the data model dictionary.
For performance reasons, this "pickle string" isn't updated when
individual properties are changed. Instead, it is only updated under
certain circumstances (saving a file, for example). As we're beginning
to use the data model for new purposes like scripting, it becomes more
important for users to be able to examine the data model. Some time ago,
a new panel was added to CellBlender called the "Data Model Browser".
This new panel essentially shows a dump of the entire data model drawn
as Blender "label" objects. However, that data model could be out of
date if any properties had been changed and the data model not updated.

This commit adds a button at the top of the Data Model Browser titled
"Regenerate Data Model" which causes the data model "pickle string" to
be regenerated from the current Blender property database. This is very
helpful for anyone writing a script and wanting to know the organization
of the data model.

A typical scenario might be a script writer who wants to create a
molecule in his or her script. The easiest way to find out what such a
molecule might contain would be to create a new molecule (in
CellBlender), and then use the Data Model Browser to view the structure
created in the data model. This will show all keys as well as the
data_model_version needed to be compatible with the current version of
CellBlender.