Added an "mcell" wrapper around the data model in the Data Model Browser
As mentioned in commit 8cc018f8, the internal data model is a dictionary
with top level keys for things like molecules, reactions, etc. However,
when the data model is exported, it is wrapped inside of an outer
dictionary under the key of "mcell". This is the format of the data
model that the user generally interacts with when importing, exporting,
and scripting with a data model. This commit simply adds that extra
"mcell" wrapping layer around the data model that is displayed in the
Data Model Browser so it will be consistent with the data model that is
imported, exported, and scripted at the user level.