Allow data model scripting of initialization without needing all keys.
The Data Model was originally planned as an internal representation of a
CellBlender model. So it was expected that all of the fields in any
CellBlender structure would be written to a data model and all of those
keys would be present in a data model that had been upgraded.

With that thought in mind, the process of copying from a Data Model to
CellBlender's internal properties assumed that all of those keys would
be present.

However, for large data model dictionaries (like initialization) this
requires that the data model contain all 30+ fields even if most of them
would be fine as defaults. So a user script that wanted to set
iterations or time_step would also have to set each of those other 30+
values.

In order to circumvent this, the "build_properties_from_data_model"
function for cellblender_initialization has been changed to check for
each key in the data model to see if it exists and only attempts to set
the corresponding property if the key exists in the data model.

There are some aspects of this solution that might need further
consideration, but this quick solution allows for easier scripting of
common models right now.