Set molecule file index to match cursor (as possible) on upgrade.
There was a problem that upgrading a data model would not show the
molecules when they existed and were refreshed. The problem was that the
molecule file index would be reset to zero and not honor the current
cursor setting. So the refresh button was working, but it was showing
the molecules at time 0 even though Blender's cursor was set to some
other value. This commit adds a small bit of code to select the file
index based on the current cursor position when possible (within bounds
of available files).