Removed globals from data model scripting and PEP8 comment
Removed globals from the exec statement in data model scripting as was
done in the previous commit for MDL scripting. This was tested and
appears to work fine so far.

Also, the PEP8 comment line was removed to see if this produces more
readable output at GitHub. Much of the CellBlender code has lines that
exceed the PEP8 length limit, and this removal is a test to see if that
flag might be being used by GitHub to format a default width.