Add Plot Output Selection Controls and Related Changes
Added the ability to select individual data output items for plotting.
The list of items to be plotted now contains an additional control to
select file generation and/or plotting. Only the selected items will
be plotted when the plot button is pressed. Since most plotters are
currently run as separate processes, this allows for making different
plots with different data from the same run and having all of them
appear on the screen at the same time.

In addition to the individual item controls, three new controls have
been added to the panel:

  Enable All   - Enables all individual items
  Disable All  - Disables all individual items
  Generate All - Generate all data regardless of individual settings

The first two are simply convenience functions to enable or disable
all items without having to do it item by item. The third (Generate
All) determines whether all of the items should generate MDL or not.
When "Generate All" is turned off, only those items that are selected
will be included in the MDL for output generation. When "Generate All"
is turned on, then all items will generate MDL for output whether each
individual item is enabled or disabled. Turning off "Generate All"
reduces the output size and may speed up the simulation. However, to
see any items that are currently disabled, the simulation would need
to be re-run with those items enabled.

The Java plotter was also updated to allow clearing of individual
plot items and to close all files so new runs can be made (which
deletes existing files) while the Java Plotting window is open.
Prior to this last change, CellBlender would produce an error when
trying to re-run a simulation when the Java plotter had a file open.

Note that at the current time, both the XmGrace and Java plotters
allow for selection of files from their menus. This allows scripting
code to generate plotter compatible files which can then be plotted
by either the XmGrace or Java plotters by opening them directly even
though those user-generated files are not specified in the plot list.
This is not currently possible with some of the other plotters. One
solution to this problem would be to provide for a command driven
plotting option where the user could bypass the automatically
generated plot list and enter the plot command manually. Because we
use a generic plot command that's translated to specific commands for
each plotter, this might work well.