Minimal and incomplete support for plotting arbitrary files.
With the addition of scripting, it would be handy to be able to add
arbitrary files (presumably created by custom scripts) to the plot file
list to be plotted along with reaction and molecule count data.

This commit includes partial support for this, however, there are some
path issues with setting and finding files from within Blender. The
plotting software generally assumes that it knows the locations of all
data files, but that assumption does not hold when the files contain
arbutrary paths. Additional work is needed to make this work smoothly.

Also, these new data items have not been added to the Data Model because
they have not been finalized yet. That is also true of the scripting
data.

This version has passed the CellBlender Test Suite.