Changed "project_files_path" function to "mcell_files_path"
The function formerly known as "project_files_path" has historically
returned a path to the mcell folder below the folder named after the
.blend file. This has made sense up to now because all files were
related to an MCell run.

However, as we're expanding the plotting capabilities, it has become
clear that there may be other files within the "project" that may not
belong directly to MCell. Examples include raw data, descriptive data,
other simulation results, and so on. The new plotting capability allows
plotting of files that may be part of the project, but may not be part
of the mcell input or output. So we will need a "project_files_path"
function, but it should return the parent of the mcell direcory rather
than the mcell directory itself.

For this reason, the "project_files_path" function has been renamed to
be "mcell_files_path" and a new "project_files_path" function will be
added.