Handle User-Specified files not in the mcell/react_data subfolder
All of the plotters assume that the file names they get are relative to
the mcell/react_data subfolder. User specified files could be located
anywhere in the file system and are given relative to the .blend file.
This change prepends the user-specified name with three sets of parent
directory references to make the file name specified in Blender relative
to the location of the reaction data expected by the plotters.