Better scripting defaults and optional data model updating
Change working directory to project area before running user script.
Also provide an option to update the data model or not. This can be
helpful when a script is just generating a data file for analysis.
In that case, there's no need to re-read the data model into properties
when the script is done.

This is particularly helpful in this version which doesn't yet include
all of the newer features in the data model. So rebuilding properties
from this version of the data model ends up effectively removing the
newer fields from the CellBlender panels.