This version stores all script files (.py and .mdl) in the data model.
More testing is needed, but this version appears to be able to save an
external data model and then recreate all properties (including scripts)
from that external data model.

Note that for some reason, the checking of the script file names in the
property panel seems to show an error until the scripts are refreshed.
This has been a problem in general which is why the refresh button was
added in the first place.

Also note that when building a simulation from a data model, the _files
subdirectory is not created until the simulation is actually run. Maybe
this should be done when the .blend file is saved so that running
scripts which build data files will have a place to put their data.
Alternatively, maybe the _files directory should be created when running
any scripts which might need it (maybe all scripts?).