Used scripting specific icons for scripting windows.
The previous use of the TEXT symbol for scripts wasn't as clear as using
the SCRIPT symbol.

Similarly, the use of the "Red Ball" for running scripts could be
confused with the same symbol used for running simulations. The
SCRIPTWIN (gear) symbol is now used for running data model scripts.