Changed "Simulation Scripting" to "Export Scripting"
In an effort to make its function clearer, I've changed the labels for
what used to be called "Simulation Scripting" to "Export Scripting"
since that's more descriptive of what's really happening. It also draws
a clearer distinction between "Export Scripting" and "Data Model
Scripting". This is still not a perfect name, but it's better than the
use of "Simulation Scripting" since Data Model Scripting also affects
any subsequent simulations.