Created subdirectories as needed for running scripts
Data Model Scripts assume that they are running in the project folder
for the .blend file that runs them. This would normally be blend_files.
This subdirectory tree is created when CellBlender runs a simulation,
but may not exist if the simulation hasn't been run. This change creates
the project subdirectory automatically when the script is run so its
default directory can be set to that location.

Additionally, this version prints the script to the console when it is
run as confirmation of what is actually running. This was added while
trying to debug a very confusing problem. The script was trying to
create an "iters" variable to take the value of the mcell iterations
parameter from CellBlender. This is done with an eval. However, the
value of that parameter was a string referencing another parameter
(which happened to be named "iters" as well). So the Python error
message was complaining that "iters" was undefined, when it was clearly
the target of the assignment. It wasn't obvious that the missing "iters"
came from the string expression used in the iterations field and not the
Python variable of the same name in the code.