Added Display of MCell stderr to Queue Runner Panel
When stderr for a task is not an empty string, the text is displayed in
the panel below the task list when that task is selected. Selecting each
task will show the stderr for that task if it's not empty.

We could add controls to show stdout as well, but that may take some
discussions since there are many ways to add such controls (in the run
panel, in the settings, ...).