Updated the data model browser to automatically copy to clipboard.
One of the problems with scripting is knowing what's in the data model
and how to copy it. The previous Blender data model browser created a
long list of labels which are not easy to navigate or to select. This
commit uses tkinter to provide a collapsable tree that automatically
copies each selected item to the clipboard for easy pasting into a
script. This is far from an ideal solution, but might be acceptable in
the near term. Anyone needing to write a script can use CellBlender to
create the objects (molecules, reactions, release sites, etc), and then
use this browser to view and copy the data model items and paste them
into their own scripts. Note that this might evolve into a Data Model
editor and that terminology is used in some of the code.

Note also that this version expects to have tkinter installed in the
Blender python directory. If it is not found, then the buttons
associated with the browser are simply not drawn. Installation of
tkinter on Linux is relatively easy. These files must be copied from
a compatible version of Python:

   python3.5/tkinter/*
      to:
        blender/2.77/python/lib/python3.5/tkinter
   python3.5/lib-dynload/_tkinter.cpython-35m-x86_64-linux-gnu.so
      to:
        blender/2.77/python/lib/python3.5/lib-dynload