Added code to generate python or text representations of data model
Clicking each item in the tree will copy a text representation of the
item that was clicked (including all sub-items) to the clipboard. This
can then be pasted into the scripting window and used as a template for
data model scripting.

The text representation can be either pure python (which ends up as a
long single line) or a text format similar to python but with nice
indentation. A good improvement would be to make the text format be
executable python as well. The selection of output mode is made from the
drop down menu at the top of the application.

This application uses the current data model contained within
CellBlender, so it will contain all of the proper "upgrade" keys so the
resulting code will produce upgradable data models. In other words, user
scripts that are based on the templates produced by this code will
themselves produce data models that can be upgraded with no changes.

As noted in an earlier commit, this feature requires installation of the
tkinter library in Blender's Python installation. If it is not present,
then the buttons to launch the browser will simply not be drawn in the
panel, and everything else should function properly.