Made use of Python's "PrettyPrint" module (pprint) for better output.
The Python pprint module produces a Python compatible data model text
that's less attractive than the previous line-by-line output. However,
since the previous line-by-line output wasn't executable Python, the
pprint module was a quick substitution that gets "reasonable" results
and can be used directly without any further editing of the code.

The purpose of this code is to provide a quick template for data model
scripting that includes all the CellBlender keys and structures.

Typical use would be as follows:

  1. Use CellBlender to build a skeleton with all needed features
  2. Open the new Data Model Browser to view the data model
  3. Click on individual structures to copy them to the clipboard
  4. Paste from the clipboard into a scripting editor
  5. The paste operation will reproduce the data model structures
  6. Write code to generate those structures for the full model

This process can be repeated for any component in the data model as
the scripting for the actual model is developed.

One of the nice features of this approach is that the data model for
each item will be fully populated with the latest CellBlender keys
and values ... including the proper data model version. This will
allow user-written scripts to take advantage of CellBlender's
built-in upgrading capability to preserve backward compatibility.