Good working prototype of Data Model Scripting
This version divides the Scripting panel into two section:

  Export Scripting
  Data Model Scripting

The "Export Scripting" section is essentially unchanged from previous
versions.

The "Data Model Scripting" is new and is further divided into:

  Make Script
  Run Script

The "Make Script" section includes controls to regenerate the Data Model
(since it's not normally regenerated until the file is saved),
optionally browse the Data Model with Tk widgets (when tkinter is
installed), select a section of the Data Model (molecules, reactions,
... or All), and finally a button to copy that section to the clipboard
as text for editing. The text copied is exactly what would be run from
the scripting window to regenerate that section.

The "Run Script" section includes a file selector and a refresh (for
some reason, files aren't automatically listed when added), a check box
to update CellBlender properties from the script (allows testing scripts
without actually updating CellBlender properties), and a button to
execute the script. The "Run Script" section also includes a selector
for choosing between internal (blend file) scripts and external (normal
file) scripts, but only Internal Scripts are implemented.

This functionality hasn't been extensively tested, but it all appears to
be working at this time.