"Clear Project" button, removal of parameters, emit handling
A "Clear Project" button was added to the scripting interface to clear
out all of the CellBlender properties in preparation for running the
script. There are times when running a script may build on the existing
properties, and there are times when it should start from scratch. This
provides the "start from scratch" option.

The removal of parameters was never implemented (not needed until now).
This commit implemented that function.

Also the "emit" value wasn't being returned from the data model into the
materials. That was added as well.