Added options to control production of Blender texts with Queue Runner
Added two controls to the Sim Runner Output/Control Options panel:

  Save Text Logs (flag to request saving of logs as Blender texts)
  Remove Task Output Texts (operator to remove "task_#_output" texts)

These control production of log texts and provide an easy way to delete
texts matching the Queue runner's naming format ("task_#_output").

Added a "make_texts" flag to the sim_runner_queue.add_task function to
control whether it creates a Blender text file ('bl_text') or assigns it
to "None". The existing code checks for "None" when determining whether
to write to a Blender text or not.