Time checking for stale runs after parameter changes
This commit adds two new time properties - one for last parameter change
and one for last run time. If the last parameter change is after the
last run time, then a warning is shown in the plotting panel.

This change is intended as an initial exploration into how we might warn
the user of "stale" data being displayed. Because this is an initial
exploration, it hasn't been fully implemented. For example, it's not
clear which panels should get the warning. Maybe the warning should be
shown in the main button panel. Maybe it should be used to show a
different icon on the "Run" button.

It's important to remember that this "stale" state will tend to be the
default when someone is working on a model, so it shouldn't be too
obnoxious.