Blender 2.77 no longer supports the "use_vertex_buffer_objects" flag
Our CellBlender Preferences panel contains a number of convenience
settings from Blender itself. Among them was use_vertex_buffer_objects.
However, in Blender 2.77 that property is no longer available. This
caused a series of "rna_uiItemR: property not found" errors whenever
that panel was displayed. This commit checks for the existence of that
property in the directory of the system object before displaying it.
This fix remains compatible with previous versions of Blender.

For reference, the Blender commit removing the vertex array option is
logged as [Bf-blender-cvs] [c402a37].