Moved "stale data" notification to main panel changed to info.
The "stale data" flag is intended to show that various display data
(plot output and visualization) is not consistent with the current state
of the model. It was originally placed in the plot data panel, but it
needed to be visible in other panels as well. Since it really applied to
the model as a whole, this change places it at the top-level CellBlender
main panel.

Originally, the "stale data" flag was shown using the bright yellow
"ERROR" icon. However, this is somewhat inappropriate because there is
no actual error when a model is being built or changed. Furthermore,
since much of the user's experience with CellBlender is spent editing
the model in various panels, it is inappropriate to be showing an error
status throughout the entire editing process. For that reason, the
"INFO" icon was chosen as a less obtrusive indicator and as being more
appropriate to the situation. The choice of ICON could be offered in the
settings area as an option.