Provided for "hidden" data model items.
Some of our data model items may not be intended for user manipulation.
An example is the "extras" subdictionary in each parameter. This commit
introduces the convention that any dictionary keys that start with an
underscore character ( _ ) will not be copied to the clipboard when the
user is generating scripting code from the data model. This is
consistent with the common Python usage of the underscore for "private"
member data.