Speed up creation of scripted objects
The creation of large numbers of scripted objects was very slow. This
commit commented out some of the selecting code and active object code
that appears to have been the problem. The selecting/active object code
was intended to leave the most recently created object as the selected
and active object. This was done by selecting and activating each object
as it was created in the loop. It appears to be much much slower to
select an object and make it active than it is to create it!! This fix
just commented out the selection/activation code. Also various print
statements were added and/or commented to help diagnose the problem.