Leave last added object as selected and active
As mentioned in the last commit message, not selecting and activating
new objects as added greatly speeds up the adding process. However, as
also mentioned, that fix did not leave the last object selected and
active. This commit adds a small amount of code to keep track of the
last added (most recent) object and then activate it at the end of the
loop. This was also done when adding surface regions to objects since
that code relies on the currently active and/or selected object as being
the target of the new region.