Fixed saving of export scripting in Data Model
The build_data_model_from_properties function for the scripting interface
includes a flag named "scripts" which is intended to trigger the saving
of the actual script text in the data model. This is useful when the
data model is being exported and should contain everything for the
model. However, that flag was being improperly used to determing whether
the CellBlender scripting properties themselves should be saved in the
data model. This change always stores the scripting properties in the
data model and only stores the script texts according to that flag.