Test of parameter sorting / change default for get/replace data model
Some test code was added to try to sort and filter parameters in the
parameter list. This is complicated by our use of internal names as the
"name" keys in the property list. The code did not work so it was
commented out. It's being left as a starting point for future work on
this item as added to the issue tracker.

The default values of the "geometry" parameter passed to both
"get_data_model" and "replace_data_model" have been changed to "False".
This speeds up the majority of scripts which do not need to read or
change the geometry of a model.