Read Partitions from Data Model
There was a bug in previous versions which kept the partition
information from being restored after an upgrade. This recreated the
model with no partitioning (the default).

The bug resulted from searching for the "partitions" key in the mcell
data model dictionary rather than in the mcell['initialization']
dictionary. The key was never found, so the partitions were not
created.