Converted start_seed and end_seed into parameters in the data model.
The CellBlender "start_seed" and "end_seed" had not been part of the
data model in the past. This caused them to be reset to 1 and 1 whenever
CellBlender was upgraded. Adding them to the data model preserves their
values across CellBlender upgrades. Additionally, these two values were
changed from integers to CellBlender parameters. This allows them to be
expressions which may include values from general parameters.