Began integrating new parameter system into existing CellBlender
The new parameter system uses a very different storage approach. Rather
than storing everything in properties, the new system has a very minimal
property footprint to hopefully improve performance. The new property
footprint for each parameter only includes 1 StringProperty along with
its implicit "Name" property. The implicit Name property holds the user
specified name in order to facilitate sorting and filtering. The only
other StringProperty is the "par_id" which holds a unique static key to
identify the parameter in the pure Python "global_params" dictionary.
As the name suggests, the "global_params" dictionary is declared as a
global variable. In the prototype, this dictionary is saved and fetched
from the .blend file via save_pre and load_post handlers. These have not
yet been ported from the prototype into CellBlender's parameter system.

This version of CellBlender adds a new optional subpanel to the
properties panel indicating that these are experimental. The subpanel
defaults to "closed" in order to minimize any confusion.