Added storage of the global_params dictionary in the .blend file.
The global_params dictionary is somewhat different from the data model,
but for now it is being stored in a similar manner. Specifically, it is
being stored as a Blender "ID Property" inside the "mcell" ID property
with the key of "global_parameters". This is still highly experimental,
but this provides some mechanism to store this non-Blender dictionary
inside the .blend file.