Added some test generation code to help with development.
Buttons were added to clear the parameter system and to generate a
series of parameters that depend on one another. These are currently
being generated as expression lists since that's what the current
system can understand. A parameter expression in the form of an
expression list looks like this:

  [ '(', 3, '+', 12, ')', '*', '3.14' ]

All of the strings in that expression list are converted as is, but the
integers are the indexes of other parameters in the parameter system. So
that expression might represent:

   ( a + b ) * 3.14

Where the index of 'a' is 3, and the index of 'b' is 12.

The plan is to write the code that evaluates the lists first, and then
add the parsing to get the lists from the expressions later. This order
was chosen to help find speed issues as soon as possible. Speed issues
would show up in the evaluation process when an expression is changed in
list form which triggers the updates of all dependencies. By contrast,
the parsing only happens once for each expression when it is entered and
is not retriggered by changes in other parameters. So parsing happens
only once per entry.

Note that the current parameter system stores expressions in a compact
string format because each expression was stored as a Blender "RNA"
property. However, the new approach is to store the entire parameter
system in a global "Pure Python" dictionary to reduce property usage.
This dictionary is then pickled into an ID property when stored in the
.blend file (very similar to how the data model is currently stored).