Added panel for sortable parameters in addition to global dictionary
The recently added parameter code was building a sortable parameter
system based on a global dictionary. This might be the fastest way to
implement parameters in Blender and uses a very small property
footprint. However, there are a considerable number of unknowns that
need to be worked out, and a fair amount of recoding needed.

This commit begins a different approach which borrows from the recently
added code to build a thin CollectionProperty that maps user parameter
names to internal ID keys. This will produce a list that is compact,
sortable, and searchable. It will use a similar update callback function
attached to the active_par_index property to draw the actual properties
that exist in the current property system. Hopefully, this should appear
identical to the existing property system.