Sortable parameters working except for deletion, and data model
The thin sortable layer (SortedParameterSystemPropertyGroup) can create
regular CellBlender parameters and display their associated properties
(expression, units, and description). Renaming a sorted parameter
changes the name of the unerlying regular parameter as it should. The
sorted parameters can be sorted or filtered (a better name might be in
order).

Deletion doesn't yet work properly, and the new parameters may not yet
be stored properly in the data model or restored from it. Similarly, all
access from the data model and the test suite API needs to be verified.

Note that the pure python dictionary code remains in this version. That
might be a faster (and better) long term solution, but it needs a lot
more work and there are many more unknowns with that approach.