Added code to build the sorting "overlay" from the data model.
There were no new items added to the data model, but code was added to
create the sorting "overlay" when reading a data model. As mentioned in
previous commit messages, parameter sorting is accomplished with a new
collection that maps user names into unique internal keys. This commit
adds code to build that mapping collection as the data model is read.