Reworked the parameter sorting without basing it on pure python code.
The previous work on parameter sorting used a new approach where the
parameter data was mostly stored in a pure python dictionary to avoid
the performance problems with Blender properties. While that is a good
approach for performance, it would require considerably more work to get
it functioning properly. As a result, this version was branched before
the new sorting code was added. Some of the changes in that other branch
were then retrofitted into this one with the goal of minimal changes.

This version appears to work with the data model and the test suite API.
As noted in earlier commits, however, not all test suite cases pass due
to Blender changing the vertex ordering of the geometrical objects that
it generates.