Added component calculation code and included in genus calculation
Added code from neuropil_tools.set_n_components to compute the number of
components within Meshalyzer. Used components to give proper genus
calculation when an object contains multiple disjoint components.

Note that the display of Meshalyzer results should probably be
overhauled. For example, the watertight property is also used to display
whether the mesh is triangulated. Similarly, certain conditions will
cause empty rows to be displayed when it might be more appropriate to
either pack the rows or show "undefined". Such an overhaul would benefit
from a brief specification of what should or should not be shown in each
of the different cases (open, non-orientable, non-manifold,
non-triangulated ...).