Added some code to try to handle circular parameter references.
This commit adds a number of features to try to handle circular
parameter references in a "least surprise" manner.

There are two kinds of circular references that might exist:

  - consistent (stable)
  - inconsistent (unstable)

A consistent circular reference exists when repeated evaluation
of each parameter in any order always gives the same result.

For example if a is 5 and b is 5, then a=b and b=a will each
evaluate to 5 ... again and again.

Here's a more complicated example:

  a = (c/2) + 1  = 5
  b = 3
  c = a + b = 8

Again, the evaluation of any parameter will return the same value
regardless of the order of evaluation. These values are not unique,
but they are consistent and stable.

Conversely, here's an inconsistent set of parameter assignments:

  a = b + 1
  b = a + 1

These are neither stable nor consistent (except possibly in the
case where they grow so large that adding 1 does not change their
numerical representation).

Because the parameter system only propagates parameter changes that
actually change a value, it will not flag consistent (stable) circular
references as being in error. It will, however, catch inconsistent
(unstable) circular references and flag them as errors.

Of course, neither can be exported to MCell as expressions because
MDL requires parameters to be defined before being used. That's not
possible when there is a circular dependency. This commit currently
recognizes that condition and exports the parameters as numerical
values when circular dependencies exist. The function:

   build_dependency_ordered_name_list()

is used to generate a parameter name list in the order needed for
MCell ... when possible. It returns "None" when that cannot be done.
A slightly updated version of that function might be what is needed
to catch consistent circular references. The expense of running that
function should also be considered.