Patch for circular reference check error in commit 7574c7a
The circular reference check counts the depth of the expression changes
and compares that to the number of parameters in the system to determine
if there is a circular reference. However, the count was being
incremented for both general parameters (capable of circular references)
and panel parameters (which cannot be referenced). So in the case where
there are many panal parameters, this count could exceed the number of
general parameters without a circular reference being present.

Note that this approach to finding circular references could be improved
by explicitly recursing through the tree rather than relying on
Blender's callback mechanism to do so.