Added more stressful parameter tests to the test suite.
Added a test to generate 100 parameters each depending on previous 3.

Added a test to generate 200 parameters each depending on previous 1.

Note that these tests generate the proper parameters, however, Python's
default maximum recursion depth of 166 is exceeded by the 200 parameter
test case during evaluation if any of the first 33 are changed. This is
currently NOT reported and should be trapped as an exception. The
difficulty is that the exception happens inside of Blender's callback
mechanism when it is calling our callback (update) function. This might
make it difficult for us to detect and handle.