Added a recursion depth estimate and increase limit to match
Python typically has a recursion depth limit of 1000. A few hundred
parameters - each depending on a few prior parameters - can exceed that
limit. This change estimates the maximum expected depth and then sets
the recursion limit based on that estimate. The increases are generous
to keep from increasing too often.

The required recursion depth can be deeper than the number of parameters
because there are several properties that depend on others within each
parameter. For example, changing an expression triggers an update that
will change the parsed expression. The change to the parsed expression
will trigger a change to a value. A change to a value will trigger
updates to other parameters, and so on.

Note that it's unclear how Blender performs the execution of update
functions. It might execute them immediately, or it might queue them for
processing after the current update has completed.