Version used for testing update call propagation for parameters.
This version was used to test update call propagation for parameters.

The commented code at the top of parameter_system.py was used as a data
model script to generate different numbers of parameters. Each parameter
depended on the previous 2 (if any) shown in the following relationships:

  ax = 1
  by = 1 + ax
  c  = 1 + ax + by
  dx = 1 + by + c
  ey = 1 + c  + dx
  f  = 1 + dx + ey
  gx = 1 + ey + f
   :   :   :    :

The total number of parameters was varied for each run, and Blender was
completely reset and reloaded twice between each run (resetting to
defaults and reloading CellBlender).

The parameter system's profiling code was used to measure the timing and
number of calls for a number of functions. The table below shows the
profiling for the "update_parameter_value" call. The "# Gen" column
shows the number of general parameters created. The "CALLS" column shows
how many calls were made to that function. Here's the table:

  FUNCTION                 # Gen       CALLS   SUM(ms)   AV(ms)
  update_parameter_value      10  =       10        0     0.00
  update_parameter_value      20  =       20      180     9.00
  update_parameter_value      30  =       30      110     3.67
  update_parameter_value      33  =       33        0     0.00
  update_parameter_value      34  <       35      300     8.57
  update_parameter_value      35  <       38        0     0.00
  update_parameter_value      40  <      119      150     1.26
  update_parameter_value      50  <    10976      680     0.06
  update_parameter_value      60  <  1346299    47400     0.04

As shown in the table, 10 parameters results in 10 calls to update the
values. 20 parameters results in 20 calls. 30 parameters gives 30 calls.
And 33 parameters gives exactly 33 calls. However, 34 parameters ends up
making 35 update calls, and it gets rapidly worse from there. Eventually
60 parameters generates over a million calls.

It's not clear if this is a result of Blender's update system or an
error in our parameter code. Given the reliable behavior below 33 calls,
it appears to be some kind of event buffer overflow which would be in
Blender. But the parameter code is very complicated, and there might be
a problem in there as well.